move descriptions to verbs folder
