+++ javascript/react/src/components/verbs/descriptions/AggregateDescription.tsx
