+++ javascript/react/src/components/verbs/descriptions/BinDescription.tsx
