+++ javascript/react/src/components/verbs/descriptions/BinarizeDescription.tsx
