+++ javascript/react/src/components/verbs/descriptions/DeriveDescription.tsx
