+++ javascript/react/src/components/verbs/descriptions/EncodeDecodeDescription.tsx
