+++ javascript/react/src/components/verbs/descriptions/EraseDescription.tsx
