+++ javascript/react/src/components/verbs/descriptions/MergeDescription.tsx
@@ -2,36 +2,40 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type { OrderbyArgs } from '@datashaper/schema'
+import type { MergeArgs } from '@datashaper/schema'
 import { memo, useMemo } from 'react'
 
 import { EMPTY_ARRAY } from '../../../empty.js'
 import type { StepDescriptionProps } from '../../../types.js'
-import { createRowEntries } from '../createRowEntries.js'
+import { createRowEntries } from '../../StepDescription/createRowEntries.js'
 import { VerbDescription } from './VerbDescription.js'
 
-export const OrderbyDescription: React.FC<StepDescriptionProps<OrderbyArgs>> =
-	memo(function OrderbyDescription(props) {
+export const MergeDescription: React.FC<StepDescriptionProps<MergeArgs>> = memo(
+	function MergeDescription(props) {
 		const rows = useMemo(() => {
 			const {
 				step: { args },
 			} = props
 			const sub = createRowEntries(
-				args.orders || EMPTY_ARRAY,
-				o => ({
-					value: o.column,
-					after: o.direction,
+				args.columns || EMPTY_ARRAY,
+				c => ({
+					value: c,
 				}),
 				1,
 				props,
 			)
 			return [
 				{
-					before: 'order',
-					value: args.orders?.length === 0 ? undefined : '',
+					before: `column${args.columns?.length !== 1 ? 's' : ''}`,
+					value: args.columns?.length === 0 ? undefined : '',
 					sub,
+				},
+				{
+					before: 'with strategy',
+					value: args.strategy,
 				},
 			]
 		}, [props])
 		return <VerbDescription {...props} rows={rows} />
-	})
+	},
+)
