+++ javascript/react/src/components/verbs/descriptions/OneHotDescription.tsx
