+++ javascript/react/src/components/verbs/descriptions/SampleDescription.tsx
