+++ javascript/react/src/components/verbs/descriptions/SpreadDescription.tsx
