+++ javascript/react/src/components/verbs/descriptions/UnfoldDescription.tsx
