+++ javascript/react/src/components/verbs/descriptions/VerbDescription.tsx
