+++ javascript/react/src/components/verbs/descriptions/WindowDescription.tsx
