+++ javascript/react/src/components/verbs/descriptions/index.ts
