+++ javascript/react/src/components/verbs/index.ts
@@ -1,1 +1,6 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './descriptions/index.js'
 export * from './renderers/index.js'
