+++ javascript/react/src/components/verbs/renderers/Filter/Filter.tsx
@@ -15,8 +15,8 @@ import {
 	useStepDataTable,
 } from '../../../../hooks/index.js'
 import type { StepComponentProps } from '../../../../types.js'
-import { LeftAlignedRow } from '../../../styles.js'
 import { FilterFunction } from '../shared/index.js'
+import { LeftAlignedRow } from '../styles.js'
 import { Container, Vertical } from './Filter.styles.js'
 
 /**
