+++ javascript/react/src/components/verbs/renderers/Rename/Rename.tsx
@@ -25,7 +25,7 @@ import {
 } from '../../../../hooks/index.js'
 import type { StepComponentProps } from '../../../../types.js'
 import { TableColumnDropdown } from '../../../controls/index.js'
-import { narrowerDropdownStyles } from '../../../styles.js'
+import { narrowerDropdownStyles } from '../styles.js'
 import {
 	useDisabled,
 	useHandleAddButtonClick,
