+++ javascript/react/src/components/verbs/renderers/SetOperation/SetOperation.hooks.tsx
@@ -8,7 +8,7 @@ import { useMemo } from 'react'
 
 import { useTableDropdownOptions } from '../../../../hooks/index.js'
 import { TableDropdown } from '../../../controls/index.js'
-import { LeftAlignedRow } from '../../../styles.js'
+import { LeftAlignedRow } from '../styles.js'
 import { icons } from './SetOperation.styles.js'
 
 export function useOthers(
