+++ javascript/react/src/components/verbs/renderers/forms/CheckboxInput.tsx
@@ -9,7 +9,7 @@ import { Fragment, memo } from 'react'
 
 import { useCheckboxChangeHandler } from '../../../../hooks/index.js'
 import type { StepChangeFunction } from '../../../../types.js'
-import { checkboxStyles } from '../../../styles.js'
+import { checkboxStyles } from '../styles.js'
 import type { CheckboxFormInput } from './CheckboxInput.types.js'
 
 export const CheckboxInput: React.FC<{
