+++ javascript/react/src/components/verbs/renderers/forms/ComboBoxInput.tsx
@@ -12,7 +12,7 @@ import {
 	useComboBoxInputValueChangeHandler,
 } from '../../../../hooks/index.js'
 import type { StepChangeFunction } from '../../../../types.js'
-import { dropdownStyles } from '../../../styles.js'
+import { dropdownStyles } from '../styles.js'
 import type { ComboBoxFormInput } from './ComboBoxInput.types.js'
 
 export const ComboBoxInput: React.FC<{
