+++ javascript/react/src/components/verbs/renderers/shared/FilterFunction.tsx
@@ -23,7 +23,7 @@ import {
 	CalendarPicker,
 	ColumnCriteriaComboBox,
 } from '../../../controls/index.js'
-import { InputExplainer, LeftAlignedRow } from '../../../styles.js'
+import { InputExplainer, LeftAlignedRow } from '../styles.js'
 import { useColumnTyping, useIsEmpty } from './FilterFunction.hooks.js'
 import {
 	BooleanToggle,
